fix: (EventService, EventServiceTest) - Foi ajustado o método de deletar um evento, e foi ajustado os testes de Event.

+++ src/main/java/br/com/ifce/easyflow/service/EventService.java
@@ -77,8 +77,11 @@ public class EventService {
 
     @Transactional
     public void delete(Long id){
-        eventRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Event not found with given id"));
+
+        if (!eventRepository.existsById(id)){
+            throw new ResourceNotFoundException("Event not found with given id.");
+        }
+
         eventRepository.deleteById(id);
     }
